+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.10.5] - 2021-12-09
+
+- Fix issues of nested iframe playback
+- Use the correct plyr way of showing Youtube player
+
 ## [6.10.4] - 2021-10-22
 
 - Fix vulnerabilities
@@ -32,14 +37,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [6.8.0] - 2021-08-11
 
 - Add an onError call back to the player
+
 ```
 const player = PlayerFactory.get(playerContainer);
 player.onError((error: BoclipsError) => {
   // do something with error
-  
+
   // remove the error element from the player
   player.getErrorHandler().clearError()
-  
+
   // to show a new one error message, you'll have to manipulate the DOM
 });
 ```